+++ server/src/server.js
@@ -181,6 +181,11 @@ class ServerInstance {
         this.connection.onCodeAction((...params) =>
             this.codeActionsProvider.onCodeActionRequest(...params)
         );
+        // Custom request from the client's extract-template command, sent
+        // after the user picked a name.
+        this.connection.onRequest("meteorToolbox/extractTemplate", (params) =>
+            this.codeActionsProvider.executeExtractTemplate(params)
+        );
         this.connection.languages.semanticTokens.on((...params) =>
             this.semanticTokensProvider.onSemanticTokensRequest(...params)
         );